requirements.txt
antlr4-python3-runtime
matplotlib
pyhton-telegram-bot
